addons/van_ban/requirements.txt
pytesseract>=0.3.0
python-docx>=0.8.11
Pillow>=8.0.0
pdfplumber>=0.5.0
web3>=5.0.0
eth-account>=0.5.0
textblob>=0.17.0
sumy>=0.8.0
google-cloud-vision>=2.0.0
azure-ai-formrecognizer>=3.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0